fix compile warning
git-svn-id: http://192.168.20.11/sequoiadb/trunk@37073 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/pd/pdComponents.cpp
@@ -84,7 +84,7 @@ UINT32 pdGetTraceComponentSize()
 
 const CHAR *pdGetTraceComponent ( UINT32 id )
 {
-   if ( (INT32)id >= pdGetTraceComponentSize() )
+   if ( id >= pdGetTraceComponentSize() )
    {
       return NULL ;
    }
